+++ JAICore/jaicore-search/src/main/java/ai/libs/jaicore/search/algorithms/standard/mcts/comparison/CosLinGammaFunction.java
@@ -54,8 +54,7 @@ public class CosLinGammaFunction implements IGammaFunction {
 			return 0.0;
 		}
 		if (visits > this.visitsToReachOne) {
-			//			g = Math.min(this.maxGamma, new AffineFunction(this.visitsToReachOne, 1, this.visitsToReachMax, this.maxGamma).apply(visits));
-			g = Math.min(this.maxGamma, Math.pow(visits - this.visitsToReachOne, 1.0/3));
+			g = Math.min(this.maxGamma, Math.pow((double)visits - this.visitsToReachOne, 1.0/3));
 		}
 		else {
 			double scaledValue = (visits - minThreshold) * 1.0 / (this.visitsToReachOne - minThreshold);
